+++ src/main/java/forex/processing/NumberFormatter.java
@@ -7,14 +7,14 @@ public class NumberFormatter {
     private NumberFormatter() {
     }
 
-    public static double round(double value) {
+    public static float round(float value) {
         return NumberFormatter.round(value, 5);
     }
 
-    public static double round(double value, int numberOfDigitsAfterDecimalPoint) {
+    public static float round(float value, int numberOfDigitsAfterDecimalPoint) {
         BigDecimal bigDecimal = new BigDecimal(value);
         bigDecimal = bigDecimal.setScale(numberOfDigitsAfterDecimalPoint, BigDecimal.ROUND_HALF_UP);
-        return bigDecimal.doubleValue();
+        return bigDecimal.floatValue();
     }
 
 }
